Bdc/feature/improve data handling (#167)
* renamed the DAL components


* Applied DAL to the main pipeline


* fix tests and module import structure


* update Pipfile.lock


* run pipenv update


* fix tests by setting database env variable


* skeleton for snapshot functionality


* save snapshots locally and to s3 after every step


* use absolute paths for local repo


* Added to check to ensure the pipeline doesn't crash when backing up a file that does not exist yet


* fix merge reviews


* fix pipfile


---------
