+++ Pipfile
@@ -10,6 +10,7 @@ name = "pypi"
 pytest = "==7.4.0"
 pre-commit = "==3.5.0"
 flake8 = "==6.0.0"
+pytest-env = "==1.0.1"
 
 [packages]
 numpy = "==1.26.1"
