+++ pytest.ini
@@ -4,3 +4,5 @@
 [pytest]
 pythonpath = src
 testpaths = tests
+env =
+    DATABASE_TYPE=Local
