+++ tests/test_bdc_pipeline.py
@@ -1,14 +1,15 @@
 # SPDX-License-Identifier: MIT
 # SPDX-FileCopyrightText: 2023 Lucca Baumgärtner <lucca.baumgaertner@fau.de>
-
+import os
 import unittest
 from unittest import mock
 
 import pandas as pd
 from pandas import DataFrame
 
-from bdc.pipeline import Pipeline, decode_s3_url
+from bdc.pipeline import Pipeline
 from bdc.steps.step import Step
+from database.leads import decode_s3_url
 
 
 class DummyStepOne(Step):
@@ -59,16 +60,21 @@ class TestPipelineFramework(unittest.TestCase):
     dummy_step_three: Step
 
     def setUp(self):
+        os.environ["DATABASE_TYPE"] = "Local"
         self.dummy_step_one = DummyStepOne()
         self.dummy_step_two = DummyStepTwo()
         self.dummy_step_three = DummyStepTwo(force_refresh=True)
         # create pipeline without actually reading from a csv file
-        with mock.patch("pandas.read_csv") as read_csv_mock:
-            self.p_one = Pipeline([self.dummy_step_one], "./not_valid.csv")
+        with mock.patch(
+            "database.leads.repository.Repository.__init__"
+        ) as init_db_mock, mock.patch(
+            "database.leads.repository.Repository.get_dataframe"
+        ):
+            init_db_mock.return_value = None
+            self.p_one = Pipeline([self.dummy_step_one], limit=10)
             self.p_two = Pipeline(
-                [self.dummy_step_two, self.dummy_step_three], "./not_valid.csv"
+                [self.dummy_step_two, self.dummy_step_three], limit=10
             )
-            read_csv_mock.assert_called_with("./not_valid.csv")
 
     def test_verify(self):
         with (
